pyproject.toml
[project]
name = "wellheard-ai"
version = "1.0.0"
description = "Voice AI Platform - Replacing Dasha.ai"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "deepgram-sdk>=3.7.0",
    "groq>=0.11.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "cartesia>=1.2.0",
    "twilio>=9.0.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "sqlalchemy>=2.0.0",
    "phonenumbers>=8.13.0",
    "python-multipart>=0.0.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest-cov>=5.0.0",
    "locust>=2.31.0",
]
